Implement savestates
